net: Use int instead of u8 for boolean flag
On some archs masking the parameter is inefficient, so don't use u8.

Reported-by: Simon Glass <sjg@chromium.org>

+++ include/net.h
@@ -178,7 +178,7 @@ void eth_halt(void);			/* stop SCC */
 const char *eth_get_name(void);		/* get name of current device */
 
 #ifdef CONFIG_MCAST_TFTP
-int eth_mcast_join(IPaddr_t mcast_addr, u8 join);
+int eth_mcast_join(IPaddr_t mcast_addr, int join);
 u32 ether_crc(size_t len, unsigned char const *p);
 #endif
